🔒 Fix predictability issue with default_rng
Replaced np.random.default_rng() with secrets.SystemRandom().gauss()
to ensure price generation is cryptographically secure.
